+++ src/ai-bundle/src/AiBundle.php
@@ -1791,6 +1791,12 @@ final class AiBundle extends AbstractBundle
                     $arguments[6] = $store['async'];
                 }
 
+                if (true === $store['hybrid_enabled']) {
+                    $arguments[] = $store['hybrid_enabled'];
+                    $arguments[] = $store['dense_vector_name'];
+                    $arguments[] = $store['sparse_vector_name'];
+                }
+
                 $definition = new Definition(QdrantStore::class);
                 $definition
                     ->setLazy(true)